Further improvements to `BotController`

+++ Core/Models/Spt/Bots/BotGenerationDetails.cs
@@ -43,13 +43,13 @@ public class BotGenerationDetails
     /// Delta of highest level of bot e.g. 50 means 50 levels above player
     /// </summary>
     [JsonPropertyName("botRelativeLevelDeltaMax")]
-    public int? BotRelativeLevelDeltaMax { get; set; }
+    public double? BotRelativeLevelDeltaMax { get; set; }
 
     /// <summary>
     /// Delta of lowest level of bot e.g. 50 means 50 levels below player
     /// </summary>
     [JsonPropertyName("botRelativeLevelDeltaMin")]
-    public int? BotRelativeLevelDeltaMin { get; set; }
+    public double? BotRelativeLevelDeltaMin { get; set; }
 
     /// <summary>
     /// How many to create and store
